pretty up OSError exceptions
as per #1, the exceptions are mighty ugly when subprocess.Popen throws
an exception. I'm not sure what all exceptions it might throw, so I'm
just going to catch OSError exceptions for now.

There are a number which will likely end up being thrown, the most
common ones probably being EACCESS and ENOENT, but I can totally see
others also coming up, for instance if the machine is out of processes,
too many open files, etc.

If we want to pretty up any of these in particular, we can, but for now
a blanket error output is fine.

fixes #1